requirements.txt
# =============================================================================
# AI EMAIL ROUTER - MINIMAL PRODUCTION DEPENDENCIES
# =============================================================================

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP client for API calls (Anthropic + Mailgun)
httpx==0.25.0

# Environment management
python-dotenv==1.0.0

# Configuration management
PyYAML==6.0.1

# Data validation
pydantic==2.5.0
email-validator==2.2.0

# Database
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Timezone handling
pytz==2023.3

# Production server
gunicorn==21.2.0

# Testing (for development)
pytest==7.4.0
pytest-asyncio==0.21.1

# CLI tools
tabulate==0.9.0
